framework/requirements.txt
httpx>=0.27.0
pyyaml>=6.0.1
jmespath>=1.0.1
playwright>=1.40.0
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
